add like, refactor
