+++ .env.example
@@ -1,0 +1,10 @@
+NODE_ENV=development
+DATABASE=
+
+JWT_SECRET=
+SESSION_SECRET=
+
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
+
+CORS_ORIGIN=http://localhost:5173
