+++ .gitignore
@@ -1,2 +1,2 @@
+**/.env
 **/node_modules
-server/.env
