--- server/models/PlaylistSchema.js
@@ -1,19 +1,0 @@
-import mongoose from "mongoose";
-
-const Schema = mongoose.Schema;
-
-const PlaylistSchema = new Schema(
-        {
-            title: String,
-            image: String,
-            url: String,
-            userId : String,
-            publicPlaylist: {
-                type: Boolean,
-                default: false
-            },
-            animeIds: [String]
-        }, { timestamps: true });
-const playlistdb = mongoose.model('playlist', PlaylistSchema);
-
-export default playlistdb;
