--- server/routes/auth.js
@@ -1,9 +1,0 @@
-
-import express from 'express';
-const router = express.Router();
-import { register, login } from '../controllers/auth.js';
-
-router.post('/register', register);
-router.post('/login', login);
-
-export default router;
