+++ utils/googleAuth.js
@@ -1,0 +1,43 @@
+import passport from "passport";
+import userdb from "../models/UserSchema.js";
+import GoogleStrategy from "passport-google-oauth20";
+
+export const configureGoogleAuth = () => {
+  passport.use(
+    new GoogleStrategy(
+      {
+        clientID: process.env.GOOGLE_CLIENT_ID,
+        clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+        callbackURL: "/auth/google/callback",
+        scope: ["email", "profile"],
+      },
+      async (accessToken, refreshToken, profile, done) => {
+        try {
+          const user = await userdb.findOne({ googleId: profile.id });
+          if (user) {
+            done(null, user);
+          } else {
+            const newUser = new userdb({
+              googleId: profile.id,
+              displayName: profile.displayName,
+              email: profile.emails[0].value,
+              image: profile.photos[0].value,
+            });
+            await newUser.save();
+            done(null, newUser);
+          }
+        } catch (err) {
+          console.log(err);
+        }
+      }
+    )
+  );
+
+  passport.serializeUser((user, done) => {
+    done(null, user);
+  });
+
+  passport.deserializeUser((user, done) => {
+    done(null, user);
+  });
+};
